[FE] feat: textArea 분리 (#50refactor)
